Try to disable watchdogs when quitting by attempting "Magic Close" feature
Some watchdogs support "Magic Close" so that watchdog will be disabled if
magic character 'V' is sent just before closing the device node.

Implement it here so that there won't be watchdog reboots when dsme is
terminated gracefully. However this has a drawback that system won't reboot
if there is a hang after graceful dsme termination.